requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
groq>=0.9.0
python-dotenv>=1.0.0
